fontmap: Cleanups
Add pango_font_map_get_resolution.
Also, make fontmap resolution a float property.

+++ pango/pango-fontmap-private.h
@@ -35,7 +35,7 @@ struct _PangoFontMap
   GHashTable *fontsets;
   GQueue fontset_cache;
 
-  double dpi;
+  float dpi;
   gboolean in_populate;
   guint serial;
 };
